@@ -84,83 +83,82 @@
     "packaging": "jar"
   },
   "dependencies": {
-    "@angular/common": "12.2.12",
-    "@angular/compiler": "12.2.12",
-    "@angular/core": "12.2.12",
-    "@angular/forms": "12.2.12",
-    "@angular/localize": "12.2.12",
-    "@angular/platform-browser": "12.2.12",
-    "@angular/platform-browser-dynamic": "12.2.12",
-    "@angular/router": "12.2.12",
-    "@fortawesome/angular-fontawesome": "0.9.0",
+    "@angular/common": "13.1.1",
+    "@angular/compiler": "13.1.1",
+    "@angular/core": "13.1.1",
+    "@angular/forms": "13.1.1",
+    "@angular/localize": "13.1.1",
+    "@angular/platform-browser": "13.1.1",
+    "@angular/platform-browser-dynamic": "13.1.1",
+    "@angular/router": "13.1.1",
+    "@fortawesome/angular-fontawesome": "0.10.1",
     "@fortawesome/fontawesome-svg-core": "1.2.36",
     "@fortawesome/free-solid-svg-icons": "5.15.4",
-    "@ng-bootstrap/ng-bootstrap": "10.0.0",
-    "@ngx-translate/core": "13.0.0",
-    "@ngx-translate/http-loader": "6.0.0",
+    "@ng-bootstrap/ng-bootstrap": "11.0.0",
+    "@ngx-translate/core": "14.0.0",
+    "@ngx-translate/http-loader": "7.0.0",
     "bootstrap": "4.6.1",
     "dayjs": "1.10.7",
     "ngx-infinite-scroll": "10.0.1",
-    "ngx-webstorage": "8.0.0",
-    "rxjs": "6.6.7",
-    "swagger-ui-dist": "4.1.2",
+    "ngx-webstorage": "9.0.0",
+    "rxjs": "7.5.1",
+    "swagger-ui-dist": "4.1.3",
     "tslib": "2.3.1",
     "zone.js": "0.11.4"
   },
   "devDependencies": {
-    "@angular-builders/custom-webpack": "12.1.3",
-    "@angular-builders/jest": "12.1.2",
-    "@angular-devkit/build-angular": "12.2.12",
-    "@angular-eslint/eslint-plugin": "12.7.0",
-    "@angular/cli": "12.2.12",
-    "@angular/compiler-cli": "12.2.12",
-    "@angular/service-worker": "12.2.12",
+    "@angular-builders/custom-webpack": "13.0.0",
+    "@angular-builders/jest": "13.0.2",
+    "@angular-devkit/build-angular": "13.1.2",
+    "@angular-eslint/eslint-plugin": "13.0.1",
+    "@angular/cli": "13.1.2",
+    "@angular/compiler-cli": "13.1.1",
+    "@angular/service-worker": "13.1.1",
     "@types/chai": "4.2.14",
     "@types/chai-string": "1.4.2",
-    "@types/jest": "27.0.3",
+    "@types/jest": "27.4.0",
     "@types/mocha": "8.2.0",
-    "@types/node": "16.11.11",
+    "@types/node": "16.11.17",
     "@types/selenium-webdriver": "4.0.11",
-    "@typescript-eslint/eslint-plugin": "5.5.0",
-    "@typescript-eslint/parser": "5.5.0",
+    "@typescript-eslint/eslint-plugin": "5.8.1",
+    "@typescript-eslint/parser": "5.8.1",
     "browser-sync": "2.27.7",
     "browser-sync-webpack-plugin": "2.3.0",
     "chai": "4.2.0",
     "chai-as-promised": "7.1.1",
     "chai-string": "1.5.0",
-    "concurrently": "6.4.0",
-    "copy-webpack-plugin": "10.0.0",
-    "eslint": "7.32.0",
+    "concurrently": "6.5.1",
+    "copy-webpack-plugin": "10.2.0",
+    "eslint": "8.5.0",
     "eslint-config-prettier": "8.3.0",
     "eslint-webpack-plugin": "3.1.1",
     "folder-hash": "4.0.1",
-    "generator-jhipster": "7.4.1",
+    "generator-jhipster": "7.5.0",
     "husky": "7.0.4",
-    "jest": "27.4.3",
+    "jest": "27.4.5",
     "jest-date-mock": "1.0.8",
     "jest-junit": "13.0.0",
-    "jest-preset-angular": "10.1.0",
+    "jest-preset-angular": "11.0.1",
     "jest-sonar-reporter": "2.0.0",
-    "lint-staged": "12.1.2",
+    "lint-staged": "12.1.4",
     "merge-jsons-webpack-plugin": "2.0.1",
     "mocha": "8.2.1",
-    "prettier": "2.5.0",
+    "prettier": "2.5.1",
     "prettier-plugin-java": "1.6.0",
     "prettier-plugin-packagejson": "2.2.15",
     "protractor": "7.0.0",
     "rimraf": "3.0.2",
-    "ts-jest": "27.0.7",
+    "ts-jest": "27.1.2",
     "ts-node": "9.1.1",
-    "typescript": "4.3.5",
+    "typescript": "4.5.4",
     "wait-on": "6.0.0",
     "webdriver-manager": "12.1.8",
-    "webpack": "5.50.0",
     "webpack-bundle-analyzer": "4.5.0",
     "webpack-merge": "5.8.0",
-    "webpack-notifier": "1.14.1"
+    "webpack-notifier": "1.15.0"
   },
   "engines": {
-    "node": ">=14.18.1"
+    "node": ">=16.13.1"
   },
   "cacheDirectories": [
     "node_modules"
@@ -168,5 +166,8 @@
   "jestSonar": {
     "reportPath": "target/test-results/jest",
     "reportFile": "TESTS-results-sonar.xml"
+  },
+  "overrides": {
+    "webpack": "5.65.0"
   }
 }
