@@ -160,7 +160,7 @@ public class OperationResource {
      */
     @GetMapping("/operations")
     public ResponseEntity<List<Operation>> getAllOperations(
-        Pageable pageable,
+        @org.springdoc.api.annotations.ParameterObject Pageable pageable,
         @RequestParam(required = false, defaultValue = "false") boolean eagerload
     ) {
         log.debug("REST request to get a page of Operations");
