@@ -50,10 +50,10 @@ export class PasswordResetFinishComponent implements OnInit, AfterViewInit {
     if (newPassword !== confirmPassword) {
       this.doNotMatch = true;
     } else {
-      this.passwordResetFinishService.save(this.key, newPassword).subscribe(
-        () => (this.success = true),
-        () => (this.error = true)
-      );
+      this.passwordResetFinishService.save(this.key, newPassword).subscribe({
+        next: () => (this.success = true),
+        error: () => (this.error = true),
+      });
     }
   }
 }
