@@ -36,10 +36,10 @@ export class PasswordComponent implements OnInit {
     if (newPassword !== this.passwordForm.get(['confirmPassword'])!.value) {
       this.doNotMatch = true;
     } else {
-      this.passwordService.save(newPassword, this.passwordForm.get(['currentPassword'])!.value).subscribe(
-        () => (this.success = true),
-        () => (this.error = true)
-      );
+      this.passwordService.save(newPassword, this.passwordForm.get(['currentPassword'])!.value).subscribe({
+        next: () => (this.success = true),
+        error: () => (this.error = true),
+      });
     }
   }
 }
