@@ -27,14 +27,14 @@ export class HealthComponent implements OnInit {
   }
 
   refresh(): void {
-    this.healthService.checkHealth().subscribe(
-      health => (this.health = health),
-      (error: HttpErrorResponse) => {
+    this.healthService.checkHealth().subscribe({
+      next: health => (this.health = health),
+      error: (error: HttpErrorResponse) => {
         if (error.status === 503) {
           this.health = error.error;
         }
-      }
-    );
+      },
+    });
   }
 
   showHealth(health: { key: string; value: HealthDetails }): void {
