@@ -30,15 +30,12 @@ export class ParseLinks {
       const url: string = section[0].replace(/<(.*)>/, '$1').trim();
       const queryString: { [key: string]: string | undefined } = {};
 
-      url.replace(
-        new RegExp('([^?=&]+)(=([^&]*))?', 'g'),
-        ($0: string, $1: string | undefined, $2: string | undefined, $3: string | undefined) => {
-          if ($1 !== undefined) {
-            queryString[$1] = $3;
-          }
-          return $3 ?? '';
+      url.replace(/([^?=&]+)(=([^&]*))?/g, ($0: string, $1: string | undefined, $2: string | undefined, $3: string | undefined) => {
+        if ($1 !== undefined) {
+          queryString[$1] = $3;
         }
-      );
+        return $3 ?? '';
+      });
 
       if (queryString.page !== undefined) {
         const name: string = section[1].replace(/rel="(.*)"/, '$1').trim();
