@@ -1,8 +1,8 @@
 jest.mock('app/core/auth/account.service');
-jest.mock('@angular/router');
 
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { Router } from '@angular/router';
+import { RouterTestingModule } from '@angular/router/testing';
 import { of, Subject } from 'rxjs';
 
 import { AccountService } from 'app/core/auth/account.service';
@@ -29,8 +29,9 @@ describe('Home Component', () => {
   beforeEach(
     waitForAsync(() => {
       TestBed.configureTestingModule({
+        imports: [RouterTestingModule.withRoutes([])],
         declarations: [HomeComponent],
-        providers: [AccountService, Router],
+        providers: [AccountService],
       })
         .overrideTemplate(HomeComponent, '')
         .compileComponents();
@@ -43,7 +44,9 @@ describe('Home Component', () => {
     mockAccountService = TestBed.inject(AccountService);
     mockAccountService.identity = jest.fn(() => of(null));
     mockAccountService.getAuthenticationState = jest.fn(() => of(null));
+
     mockRouter = TestBed.inject(Router);
+    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
   });
 
   describe('ngOnInit', () => {
