@@ -81,7 +81,7 @@ describe('HasAnyAuthorityDirective tests', () => {
       mockAccountService.hasAnyAuthority = jest.fn(() => false);
 
       // WHEN
-      authenticationState.next();
+      authenticationState.next(null);
       fixture.detectChanges();
 
       // THEN
@@ -91,7 +91,7 @@ describe('HasAnyAuthorityDirective tests', () => {
       mockAccountService.hasAnyAuthority = jest.fn(() => true);
 
       // WHEN
-      authenticationState.next();
+      authenticationState.next(null);
       fixture.detectChanges();
 
       // THEN
@@ -115,7 +115,7 @@ describe('HasAnyAuthorityDirective tests', () => {
 
       // WHEN
       jest.clearAllMocks();
-      authenticationState.next();
+      authenticationState.next(null);
 
       // THEN
       expect(mockAccountService.hasAnyAuthority).toHaveBeenCalled();
@@ -123,7 +123,7 @@ describe('HasAnyAuthorityDirective tests', () => {
       // WHEN
       jest.clearAllMocks();
       hasAnyAuthorityDirective.ngOnDestroy();
-      authenticationState.next();
+      authenticationState.next(null);
 
       // THEN
       expect(mockAccountService.hasAnyAuthority).not.toHaveBeenCalled();
