@@ -51,21 +51,21 @@ class TokenProviderSecurityMetersTests {
     }
 
     @Test
-    public void testValidTokenShouldNotCountAnything() {
+    void testValidTokenShouldNotCountAnything() {
         Collection<Counter> counters = meterRegistry.find(INVALID_TOKENS_METER_EXPECTED_NAME).counters();
 
-        assertThat(aggregate(counters)).isEqualTo(0);
+        assertThat(aggregate(counters)).isZero();
 
         String validToken = createValidToken();
 
         tokenProvider.validateToken(validToken);
 
-        assertThat(aggregate(counters)).isEqualTo(0);
+        assertThat(aggregate(counters)).isZero();
     }
 
     @Test
-    public void testTokenExpiredCount() {
-        assertThat(meterRegistry.get(INVALID_TOKENS_METER_EXPECTED_NAME).tag("cause", "expired").counter().count()).isEqualTo(0);
+    void testTokenExpiredCount() {
+        assertThat(meterRegistry.get(INVALID_TOKENS_METER_EXPECTED_NAME).tag("cause", "expired").counter().count()).isZero();
 
         String expiredToken = createExpiredToken();
 
@@ -76,7 +76,7 @@ class TokenProviderSecurityMetersTests {
 
     @Test
     void testTokenUnsupportedCount() {
-        assertThat(meterRegistry.get(INVALID_TOKENS_METER_EXPECTED_NAME).tag("cause", "unsupported").counter().count()).isEqualTo(0);
+        assertThat(meterRegistry.get(INVALID_TOKENS_METER_EXPECTED_NAME).tag("cause", "unsupported").counter().count()).isZero();
 
         String unsupportedToken = createUnsupportedToken();
 
@@ -86,8 +86,8 @@ class TokenProviderSecurityMetersTests {
     }
 
     @Test
-    public void testTokenSignatureInvalidCount() {
-        assertThat(meterRegistry.get(INVALID_TOKENS_METER_EXPECTED_NAME).tag("cause", "invalid-signature").counter().count()).isEqualTo(0);
+    void testTokenSignatureInvalidCount() {
+        assertThat(meterRegistry.get(INVALID_TOKENS_METER_EXPECTED_NAME).tag("cause", "invalid-signature").counter().count()).isZero();
 
         String tokenWithDifferentSignature = createTokenWithDifferentSignature();
 
@@ -97,8 +97,8 @@ class TokenProviderSecurityMetersTests {
     }
 
     @Test
-    public void testTokenMalformedCount() {
-        assertThat(meterRegistry.get(INVALID_TOKENS_METER_EXPECTED_NAME).tag("cause", "malformed").counter().count()).isEqualTo(0);
+    void testTokenMalformedCount() {
+        assertThat(meterRegistry.get(INVALID_TOKENS_METER_EXPECTED_NAME).tag("cause", "malformed").counter().count()).isZero();
 
         String malformedToken = createMalformedToken();
 
