@@ -1,17 +1,7 @@
 function setupProxy({ tls }) {
   const conf = [
     {
-      context: [
-        '/api',
-        '/services',
-        '/management',
-        '/swagger-resources',
-        '/v2/api-docs',
-        '/v3/api-docs',
-        '/h2-console',
-        '/auth',
-        '/health',
-      ],
+      context: ['/api', '/services', '/management', '/v3/api-docs', '/h2-console', '/auth', '/health'],
       target: `http${tls ? 's' : ''}://localhost:8080`,
       secure: false,
       changeOrigin: tls,
